pyproject.toml
[project]
name = "toxicity-service"
version = "0.1.0"
description = "Toxicity service"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "grpcio>=1.78.0",
    "grpcio-tools>=1.78.0",
    "protobuf>=6.33.5",
    "torch>=2.10.0",
    "transformers>=5.3.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
  "pytest==8.3.3",
  "httpx>=0.27.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
